pyproject.toml
[project]
name = "TiebaManageBot"
version = "1.4.0"
description = "TiebaManageBot"
authors = [{ name = "mxcoras", email = "mxcoras@outlook.com" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "aiotieba>=4.6.1",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "cashews[diskcache]>=7.4.4",
    "httpx>=0.28.1",
    "nb-cli>=1.6.0",
    "nonebot-adapter-onebot>=2.4.6",
    "nonebot-plugin-alconna>=0.60.3",
    "nonebot-plugin-apscheduler>=0.5.0",
    "nonebot2[fastapi]>=2.4.4",
    "pyjwt>=2.10.1",
    "python-multipart>=0.0.21",
    "redis[hiredis]>=6.4.0",
    "sqlalchemy>=2.0.45",
    "tiebameow[orm,renderer]>=0.2.4,<0.3.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]

[tool.nonebot]
plugin_dirs = ["src/plugins"]
builtin_plugins = []

[tool.nonebot.plugins]
"@local" = []
nonebot-plugin-apscheduler = ["nonebot_plugin_apscheduler"]
nonebot-plugin-alconna = ["nonebot_plugin_alconna"]


[tool.nonebot.adapters]
"@local" = []
nonebot-adapter-onebot = [{name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11"}]
